Migrate map-world_countries from yaml to json
Since limn does no longer allow to directly fetch yaml files, we had
to adapt geowiki monitoring to fetch the map-world_countries
datasource as json. However, that is a problem as running the
monitoring against local checkouts of the data repositiories, which
still only contain the yaml and no json representation. Hence, we
update the yaml datasources to json, so monitoring runs again with or
without local checkouts of the data reposititories.

Change-Id: I5c1191eb35a02a0cabe803e1e57e8856ebf5c7ac
Card: Analytics 1253